create directory "widget" for each Screen in "authentication" and "home", Rename content in "dashboard" directory into snake style Create appbar.dart for dashboard.dart Resize popular_courses_widget.dart Component in dashboard.dart Adding images & strings for popular_courses_widget.dart
