+++ lib/src/constants/image_strings.dart
@@ -37,8 +37,9 @@ const String icGraphicDesign = "assets/icons/ic_graphic_design.png";
 const String icArtificialIntelligence = "assets/icons/ic_artificial_intelligence.png";
 
 // Courses
-const String imgCoursePython = "assets/images/courses/img_introduction_to_python.png";
-const String imgCourseFlutter = "assets/images/courses/img_introduction_to_flutter.png";
+const String imgCoursePython = "assets/images/courses/img_introduction_to_python.jpg";
+const String imgCourseFlutter = "assets/images/courses/img_introduction_to_flutter.jpg";
+const String imgCourseJava = "assets/images/courses/img_introduction_to_java.jpeg";
 const String imgCourseArchitecture = "assets/images/courses/img_introduction_to_architecture_design.png";
 const String imgCourseCoding = "assets/images/courses/img_coding.png";
 const String imgCourseDesigning = "assets/images/courses/img_designing.png";
