+++ lib/src/constants/sizes.dart
@@ -1,6 +1,6 @@
 import 'dart:ui';
 
-
+// Text Sizes
 const double txtHeadlineMediumSize = 27.0;
 const double txtHeadlineSmallSize = 24.0;
 
@@ -14,15 +14,22 @@ const double txtBodyLargeSize = 15.0;
 const double txtBodyMediumSize = 12.0;
 const double txtBodySmallSize = 11.0;
 
+// Default Sizes
 const double txtDefaultSize = 17.0;
+const double defaultPaddingSize = 8.0;
+const double defaultHeight = 50.0;
 
-const double imgSplashScreenBlobSize = 150.0;
-
+// Button Sizes
 const double btnWidth = 250.0;
 const double btnHeight = 50.0;
 
+// Background Sizes
 const double bgWidth = 350.0;
 const double bgHeight = 720.0;
 
-const double defaultPaddingSize = 8.0;
-const double defaultHeight = 50.0;
+
+// Items Sizes
+const double imgSplashScreenBlobSize = 150.0;
+const double coursesSizeWidth = 220;
+const double coursesSizeHeight = 150;
+const double coursesSize = coursesSizeHeight*2;
